Move exception type tests to own file
