+++ analyzers/tests/SonarAnalyzer.UnitTest/SymbolicExecution/Roslyn/RoslynSymbolicExecutionTest.TryCatchFinally.cs
@@ -897,223 +897,6 @@ tag = ""End"";";
             validator.TagStates("End").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
         }
 
-        [TestMethod]
-        public void Exception_FieldReference()
-        {
-            const string code = @"
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw fieldException;
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(1).And.ContainSingle(x => HasExceptionOfType(x, "NotImplementedException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_LocalReference()
-        {
-            const string code = @"
-NotImplementedException exception = null;
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw exception;
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(1).And.ContainSingle(x => HasExceptionOfType(x, "NotImplementedException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_PropertyReference()
-        {
-            const string code = @"
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw PropertyException;
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(1).And.ContainSingle(x => HasExceptionOfType(x, "NotImplementedException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_ParameterReference()
-        {
-            const string code = @"
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw ex;
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code, ", NotImplementedException ex").Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(1).And.ContainSingle(x => HasExceptionOfType(x, "NotImplementedException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_ArrayElementReference()
-        {
-            const string code = @"
-NotImplementedException[] exceptions = null;
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw exceptions[0];
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(2)
-                     .And.ContainSingle(x => HasExceptionOfType(x, "NotImplementedException"))
-                     .And.ContainSingle(x => HasExceptionOfType(x, "IndexOutOfRangeException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_MethodInvocation()
-        {
-            const string code = @"
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw CreateException();
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch", // In case CreateException throws
-                "InCatch", // throw
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(2)
-                     .And.ContainSingle(x => HasExceptionOfType(x, "NotImplementedException"))
-                     .And.ContainSingle(x => HasUnknownException(x));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_Conversion()
-        {
-            const string code = @"
-var tag = ""BeforeTry"";
-var obj = new ArgumentNullException();
-try
-{
-    tag = ""InTry"";
-    throw obj;
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(1).And.ContainSingle(x => HasExceptionOfType(x, "ArgumentNullException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
-        [TestMethod]
-        public void Exception_MultipleConversion()
-        {
-            const string code = @"
-var tag = ""BeforeTry"";
-try
-{
-    tag = ""InTry"";
-    throw (System.IO.IOException) new System.IO.FileNotFoundException();
-}
-catch
-{
-    tag = ""InCatch"";
-}
-tag = ""AfterCatch"";";
-            var validator = SETestContext.CreateCS(code).Validator;
-            validator.ValidateTagOrder(
-                "BeforeTry",
-                "InTry",
-                "InCatch",
-                "AfterCatch");
-
-            validator.TagStates("InCatch").Should().HaveCount(1).And.ContainSingle(x => HasExceptionOfType(x, "FileNotFoundException"));
-            validator.TagStates("AfterCatch").Should().HaveCount(1).And.ContainSingle(x => HasNoException(x));
-        }
-
         private static bool HasNoException(ProgramState state) =>
             state.Exception == null;
